Solved another bug

+++ merged_model.py
@@ -202,7 +202,7 @@ class DCGAN_Caps(object):
         if could_load:
             counter = checkpoint_counter
             print(" [*] Load SUCCESS")
-            start_epoch = (counter / 1094)
+            start_epoch = int(counter / 1094)
         else:
             print(" [!] Load failed...")
